Timer supports 32 bit timekeeping

+++ timer.c
@@ -4,7 +4,7 @@
 #include "uart.h"
 
 volatile uint16_t sys_time = 0;
-volatile uint16_t sys_time_sec_low = 0;
+volatile uint32_t sys_time_sec_low = 0;
 //volatile uint8_t sys_time_sec_high = 0;
 
 char time[SYSTIME_STR_LEN];
@@ -34,14 +34,14 @@ static uint16_t get_sys_time_ms(void)
     return sys_time % 1000;
 }
 
-static uint16_t get_sys_time_s(void)
+static uint32_t get_sys_time_s(void)
 {
     return sys_time_sec_low;
 }
 
 char * get_timestamp_str()
 {
-    sprintf(time, "[%6u.%03u] ", get_sys_time_s(), get_sys_time_ms());
+    sprintf(time, "[%6lu.%03u] ", get_sys_time_s(), get_sys_time_ms());
     return time;
 }
